pyproject.toml
[project]
name = "smart-crawler"
version = "0.0.0"
description = "A retrieval API with an agent inside. Built on crawl4ai."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.8.5,<0.9",
    "pydantic>=2.7",
    "rapidfuzz>=3.0",
    "tiktoken>=0.12",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "openai>=1.30",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
benchmark = [
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "numpy>=1.26",
    "openai>=1.30",
    "tavily-python>=0.5",
    "python-dotenv>=1.0",
]
dev = [
    # tests
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov",
    "pytest-recording",   # VCR cassettes for crawl4ai calls
    "hypothesis",         # property-based fuzzing of extractors
    "syrupy",             # snapshot tests for extracted outputs
    # lint / format / types / security
    "ruff>=0.8",
    "basedpyright",
    "semgrep",
    "bandit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E", "F", "W",   # pycodestyle + pyflakes
    "I",             # isort
    "B",             # bugbear
    "UP",            # pyupgrade
    "SIM",           # simplify
    "RUF",           # ruff-specific
    "S",             # bandit-lite
    "ASYNC",         # async pitfalls
]
ignore = ["S101"]  # allow assert in tests

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S", "B"]

[tool.basedpyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["smart_crawler", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
